tabs, create post button, edit mode button, grid spacing, changing detail divs on click
